encode/decode: refactor into new module bit_encoding
This is pretty-much a move-only commit.
