+++ src/bit_encoding/bititer.rs
