+++ src/bit_encoding/bitwriter.rs
