+++ src/lib.rs
@@ -34,13 +34,10 @@ pub use elements_miniscript as miniscript;
 mod macros;
 
 mod analysis;
+pub mod bit_encoding;
 pub mod bit_machine;
-pub mod bititer;
-pub mod bitwriter;
 pub mod core;
 pub mod dag;
-mod decode;
-mod encode;
 pub mod jet;
 pub mod merkle;
 #[cfg(feature = "elements")]
@@ -49,6 +46,11 @@ pub mod types;
 // #[cfg(test)]
 // mod test_progs;
 mod util;
+
+pub use bit_encoding::bititer;
+pub use bit_encoding::bitwriter;
+use bit_encoding::decode;
+use bit_encoding::encode;
 
 #[cfg(feature = "elements")]
 pub use crate::policy::Policy;
